tests: move packager back in scripts
