+++ scripts/packager/utils.mjs
@@ -1,20 +1,34 @@
+// @ts-check
+
 import { execSync } from "child_process";
 import { resolve } from "path";
 import { existsSync, readdirSync, unlinkSync } from "fs";
 import { ARTIFACT_PATTERNS } from "./constants.mjs";
 
-/** Returns true if `value` is a 32-character hexadecimal string. */
+/**
+ * Returns true if `value` is a 32-character hexadecimal string.
+ * @param {string} value
+ * @returns {boolean}
+ */
 export function isValidHexKey(value) {
   return /^[0-9a-fA-F]{32}$/.test(value);
 }
 
-/** Returns true if `value` is a positive integer. */
+/**
+ * Returns true if `value` is a positive integer.
+ * @param {string} value
+ * @returns {boolean}
+ */
 export function isPositiveInteger(value) {
   const n = Number(value);
   return Number.isInteger(n) && n > 0;
 }
 
-/** Returns true if `value` is a valid TCP/UDP port number (1–65535). */
+/**
+ * Returns true if `value` is a valid TCP/UDP port number (1–65535).
+ * @param {string} value
+ * @returns {boolean}
+ */
 export function isValidPort(value) {
   const n = Number(value);
   return Number.isInteger(n) && n >= 1 && n <= 65535;
@@ -44,7 +58,11 @@ export function commandExists(cmd) {
   }
 }
 
-/** Returns true if `filename` matches any known artifact pattern. */
+/**
+ * Returns true if `filename` matches any known artifact pattern.
+ * @param {string} filename
+ * @returns {boolean}
+ */
 export function isArtifact(filename) {
   return ARTIFACT_PATTERNS.some((re) => re.test(filename));
 }
@@ -78,7 +96,8 @@ export function cleanupMediaFiles(outputDir) {
       }
     }
   } catch (err) {
-    console.warn(`Warning: error during media cleanup: ${err.message}\n`);
+    const message = err instanceof Error ? err.message : "Unknown Error";
+    console.warn(`Warning: error during media cleanup: ${message}\n`);
   }
   console.log("Media files cleanup completed.");
 }
