+++ tests/integration/scenarios/dash_multi-track.test.js
@@ -1,6 +1,6 @@
 import { describe, beforeEach, afterEach, it, expect } from "vitest";
 import RxPlayer from "../../../dist/es2017";
-import { multiAdaptationSetsInfos } from "../../contents/DASH_static_SegmentTimeline";
+import { multiAdaptationSetsInfos } from "../../contents/static/DASH_static_SegmentTimeline";
 import { checkAfterSleepWithBackoff } from "../../utils/checkAfterSleepWithBackoff.js";
 import sleep from "../../utils/sleep.js";
 import launchEventInterruptionTests from "../utils/stop_on_event";
