+++ tests/integration/scenarios/directfile.test.js
@@ -16,7 +16,7 @@
 
 import { describe, beforeEach, afterEach, it, expect } from "vitest";
 import RxPlayer from "../../../dist/es2017";
-import directfileInfos from "../../contents/directfile_webm";
+import directfileInfos from "../../contents/static/directfile_webm";
 import sleep from "../../utils/sleep.js";
 import { waitForLoadedStateAfterLoadVideo } from "../../utils/waitForPlayerState";
 import { checkAfterSleepWithBackoff } from "../../utils/checkAfterSleepWithBackoff.js";
