+++ tests/integration/scenarios/fast-switching.test.js
@@ -1,5 +1,5 @@
 import { describe, beforeEach, afterEach, it, expect } from "vitest";
-import { manifestInfos } from "../../contents/DASH_static_SegmentTimeline";
+import { manifestInfos } from "../../contents/static/DASH_static_SegmentTimeline";
 import RxPlayer from "../../../dist/es2017";
 import sleep from "../../utils/sleep.js";
 import { waitForLoadedStateAfterLoadVideo } from "../../utils/waitForPlayerState";
