+++ tests/integration/scenarios/loadVideo_options.test.js
@@ -19,7 +19,7 @@ import RxPlayer from "../../../dist/es2017";
 import { MULTI_THREAD } from "../../../dist/es2017/experimental/features/index.js";
 import { EMBEDDED_DASH_WASM } from "../../../dist/es2017/__GENERATED_CODE/index.js";
 import TestWorkerEmbed from "../../embedded_worker_bundle";
-import { manifestInfos } from "../../contents/DASH_static_SegmentTimeline";
+import { manifestInfos } from "../../contents/static/DASH_static_SegmentTimeline";
 import sleep from "../../utils/sleep.js";
 import {
   /* waitForState, */ waitForLoadedStateAfterLoadVideo,
